Move the egui-backed EguiTextMeasurer into a dedicated gfx module
