+++ crates/gfx/Cargo.toml
@@ -9,3 +9,5 @@ egui = { workspace = true }
 egui-winit = { workspace = true }
 egui-wgpu = { workspace = true }
 pollster = { workspace = true }
+css = { path = "../css" }
+layout = { path = "../layout" }
